requirements.txt
tensorflow>=2.13
numpy>=1.24
pandas>=2.0
Pillow>=10.0
scikit-learn>=1.3
streamlit>=1.30
tqdm>=4.66
transformers>=4.38
torch>=2.1
fastapi>=0.111
uvicorn>=0.30
python-multipart>=0.0.9
